Rendre la garde de base utilisable hors de vitest
db-guard importe describe, donc un script tsx qui appelle assertDisposableTestDb()
échouait au chargement.
